requirements.txt
colorama
types-colorama

prettytable

pydantic
pydantic-settings

GitPython
